requirements.txt
django>=6.0
djangorestframework>=3.16
django-cors-headers>=4.9
gunicorn>=23.0
psycopg2-binary>=2.9
dj-database-url>=2.3
whitenoise>=6.8
python-decouple>=3.8
